refactor: 🛠  moved messageRequest.ts in types directory
