+++ src/contentScript/frameScript.tsx
@@ -1,6 +1,6 @@
 import { rootRender } from '@utils/render';
 import { removeFromStorage } from '@utils/storage';
-import { MessageRequest } from '@utils/MessageRequest';
+import { MessageRequest } from '@utils/types/MessageRequest';
 import '@styles/tailwind.css';
 // ICONS
 import Close from '@icons/X.svg';
